broadcast if defined
